extract package walk logic into utility
